Quickfix for trivia
Pending my getting shanai onto a stable host with other people having access to restart as well as completing the v2 port, this will make approving trivia questions somewhat acceptable again.

/accept no longer immediately saves the approved question to the trivia database. This may cause lost questions in some cases, but will remove some server lag caused by approving questions.

Additionally we add a new command /savedb which is used periodically after approving a batch of trivia questions. This command will lag the server.

As a result we essentially exchange some data reliability for speed. So long as the server does not crash while trivia folks are approving questions, nothing will be lost. What we gain is the ability to approve questions without lagging the server for 2-4 seconds every time.