unbanning micle berry (released)
The WiFi tournament, the All Star Cup, has been fully detailed by the official site. Like we reported before, this tournament is open only to Black 2 & White 2 users and has you participate in Double Battles and use any of the Pokémon in the new Unova Pokédex from Victini to Genesect, with no Pokémon restrictions. Registration for this tournament opens on September 14th and lasts until September 20th while the tournament itself runs from September 21st to September 23rd. Like The Beginning Cup, it is seperated into three divisions based on your age; Junior, Senior & Master with 50,000 places.

Unique to this cup is a special present. If you register for the tournament, you will be given the rare Micle Berry which is yet to be available in Generation V. You are given it in the Dream World so you can grow more at will.