Mafia: Community Suggested Features
Along with some of my own, of course. :P

- Creates the directory "mafiathemes" on PO > 2.0.05 to fix some bugs
(if it already doesn't exist)
- __role__.hide: Can hide a role from /roles and /sides
- __side__.hide: Can hide a side from /sides
- __global__.dayStart: Starts at standby, not the night (might be buggy,
someone should review it since i'm not the best in coding for the mafia
script)
- __global__.noSummaryBot: Removes the "Game:" part from the summary so
themers can insert linebreaks (\n), tabs (\t) etc. to make it more
readable.
- __role__.lynch.convertRoles: Converts someone else when __role__ gets
lynched (syntax same as onDeath.convertRoles).
- __role__.lynch.convertRolesmsg: Custom message for lynch.convertRoles

I also added Line Numbers to the state switcher so it will now also
display the line number in which the error was.