Hangman: Auto-Games when channel is idling
Also added small stuff like a line confirming the answer to the game you started (to check spelling and other mistakes).

Auto-Games are based on BeastCharizard's idea, it automatically starts a Hangman game if no one starts one after 4 minutes (including the 60 seconds the last winner has to start it).

At line 23, a file with the answer must be defined. I'm not sure where it would be the best place to put the file, so I left in my Dropbox for now. Change to a better place later.