Hangman: Event Games
Adding Event Games, which are Auto-Games that give Leaderboard points, as requested.
Still need some stuff like being able to remove entries from the question bank.
Also, once this is updated, the Auto-Games will not work until some HA use /addquest to add questions to the bank. Another option is to copy the autogames file from my dropbox and save it on the server as "hangmanq.txt".