Sand Rush Ban + More consistent Tier Check messages
Also groups related bans together, like putting all the ability bans in
the same general area, making it easier to find existing bans.

Consistency in checks regards always telling tier when applicable,
especially on Drought Ban which always said "Smogon UU" even when
changing to "Monotype". Weather related abilities all had different
messages it seemed too, so better consistency there as well.

Half the messages said something along the lines of "change it in the
teambuilder" and half didn't. I opted to remove this section because it
just makes the message more cluttered. If you want, I can add it in
again at the end of all applicable tiers. When you try to join a tier
with a banned pokemon it doesn't say to change it ((19:08:05) The
Pokemon 'Mewtwo' is banned on tier 'ORAS NU' for the following reasons:
) so that's why I opted to remove the instances.