Fixed the issue where enabletours and disabletours don't exist.
I had difficulty with this and I'm not fully sure how this section works. I just hope I didn't break anything that I'm not aware of but I've tested and it works as it should now. Before, if /enabletours was already enabled server responds command doesn't exist. Same with /disabletours when they were already disabled.