channelcommands.js /register security update + moved commands (1/2)
* Added security to /register.
* Moved topic related commands from usercommands.js to here since they in wrong script.

/register before allows anyone, even if they are not part of the cauth, to get c-owner of the channel when there are no other c-owners in it. Even if there already still have c-admins, c-mods and c-members, anyone can just come in and /register for c-owner.

This update adds a check so that if there are no c-owners but c-admins, only a c-admin will be able to /register. Same if there are no c-owners and c-admins, a c-mod can. Even if no c-owners, c-admins and c-mods, only a c-member be able to /register. If there no c-owners, c-admins, c-mods and c-members, then anyone can /register.