Pyramid HazardRoom
This is a change I suggested in the thread some months ago, then eventually tested on another server in a heavily modified version of pyramid while PO's pyramid was down. 

The changes are:
- Exactly 3 Hazards per Hazard Room
- Hazard Rooms tells you what the 3 Hazards are 
- Added Spider Web (Rapid Spin or Struggle Bug to remove) - adds difficulty because otherwise this update makes Hazard Rooms way easier, and also makes some more unique mons more viable like Flygon or Delibird.

Pros:
- Requires a lot more fast communication between team members. Like all 3 of you have to commit to clearing a different hazard, or all of you are doomed.
- Also rewards knowing not only what your own party can do, but having an idea of what the mons in your teammates' parties are capable of clearing.
- No more annoying "You cleared 6% of the Stream" and "you all died to 90% of invisible Electric Fence that wasn't there"
- It was lots of fun when testing

Cons:
I suppose you're really screwed when you're the last player alive on your team, but that's true for most any room. 

This is tested, but not with the newer version of pyramid, so if there are any changed functions that it calls they will need to be updated. (I scanned through to see if there were but it's hard to catch these things.)