Allow for rock to serve as $25 pawnable
One aspect of the game that acts as a limiting factor on the overall kindness of the safari community is the excessive distribution of rocks.  Players are given extremely high numbers of rocks, considering gacha, itemfinder, login bonus, and contest prizes.  The only thing players can do with these is to rock other players — an almost derogatory action.  This would make since to be able to do for $50, and maybe given seldom, i.e. as a contest prize only, but the rate at which Safari gives them encourages this derogatory action.  Players should have an incentive to build each other up not down, so I reclassified rocks as pawnables that can be pawned for $25-$27 using /pawn rock:1, or traded: just like any other pawnable.