Update Android Links?
According to @MidwayMarshall the PC links are supposed to work for current versions of Android. However, there seem to be complications all over the place, so I am not deleting anything we don't have to. I have had an Android player test links on three different versions (including the newest) which various success and it's hard to tell what works and what doesn't. For now I'm just going to send all the PC links and none of the old Android links because that's what's "supposed" to work. This might break everything so I might need to revert.